+++ interpreter/include/process_ram_kb.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <fstream>
+#include <string>
+#include <iostream>
+
+long get_process_ram_kb();
+
+void print_system_info(bool enabled);