+++ interpreter/include/shell_utilities.h
@@ -1,0 +1,6 @@
+#prama once
+#include <iostream>
+#include <termios.h>
+
+void set_buffered_input(bool enable)
+void clear_screen()